pyproject.toml
[tool.poetry]
name = "evetrade-historic-volume"
version = "0.1.0"
description = "Imputes the historical trading volume for the EVE Online Universe"
authors = ["awhipp"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
boto3 = "^1.26.21"
aiohttp = "^3.8.3"
asyncio = "^3.4.3"
redis = "4.3.4"
setuptools = "^67.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
